Reduce solver template compile pressure
