+++ src/meson.build
@@ -1,4 +1,11 @@
-ab_sources = ['acacia-bonsai.cc']
+ab_sources = [
+  'acacia-bonsai.cc',
+  'solver/solver_invoker.cc',
+  'solver/solve_game.cc',
+  'solver/solve_game_array_bitset.cc',
+  'solver/solve_game_vector_bitset.cc',
+  'solver/solve_game_vector_boolvec.cc',
+]
 
 if get_option('build_python')
   subdir ('python')
@@ -10,4 +17,3 @@ ab_exe = executable ('acacia-bonsai', ab_sources,
                      install: true,
                      install_tag: 'cli',
                      dependencies : [posets_dep, spot_dep, bddx_dep, stdsimd_dep])
-
