+++ src/solver/configured_components.hh
@@ -1,0 +1,73 @@
+#pragma once
+
+#include "configuration.hh"
+
+#if ACACIA_ENABLE_AUT_PREPROCESSOR_NO_PREPROCESSING
+#include "aut_preprocessors/no_preprocessing.hh"
+#endif
+
+#if ACACIA_ENABLE_AUT_PREPROCESSOR_STANDARD
+#include "aut_preprocessors/standard.hh"
+#endif
+
+#if ACACIA_ENABLE_AUT_PREPROCESSOR_SURELY_LOSING
+#include "aut_preprocessors/surely_losing.hh"
+#endif
+
+#if ACACIA_ENABLE_AUT_PREPROCESSOR_ELEVATOR
+#include "aut_preprocessors/elevator.hh"
+#endif
+
+#if ACACIA_ENABLE_BOOLEAN_STATES_FORWARD_SATURATION
+#include "boolean_states/forward_saturation.hh"
+#endif
+
+#if ACACIA_ENABLE_BOOLEAN_STATES_NO_BOOLEAN_STATES
+#include "boolean_states/no_boolean_states.hh"
+#endif
+
+#include "actioners/direction.hh"
+
+#if ACACIA_ENABLE_ACTIONER_NO_IOS_PRECOMPUTATION
+#include "actioners/no_ios_precomputation.hh"
+#endif
+
+#if ACACIA_ENABLE_ACTIONER_STANDARD
+#include "actioners/standard.hh"
+#endif
+
+#if ACACIA_ENABLE_IOS_PRECOMPUTER_DELEGATE
+#include "ios_precomputers/delegate.hh"
+#endif
+
+#if ACACIA_ENABLE_IOS_PRECOMPUTER_FAKE_VARS
+#include "ios_precomputers/fake_vars.hh"
+#endif
+
+#if ACACIA_ENABLE_IOS_PRECOMPUTER_POWSET
+#include "ios_precomputers/powset.hh"
+#endif
+
+#if ACACIA_ENABLE_IOS_PRECOMPUTER_STANDARD
+#include "ios_precomputers/standard.hh"
+#endif
+
+#if ACACIA_ENABLE_IOS_PRECOMPUTER_MONA
+#include "ios_precomputers/mona.hh"
+#endif
+
+#if ACACIA_ENABLE_INPUT_PICKER_CRITICAL
+#include "input_pickers/critical.hh"
+#endif
+
+#if ACACIA_ENABLE_INPUT_PICKER_CRITICAL_FULLRND
+#include "input_pickers/critical_fullrnd.hh"
+#endif
+
+#if ACACIA_ENABLE_INPUT_PICKER_CRITICAL_PQ
+#include "input_pickers/critical_pq.hh"
+#endif
+
+#if ACACIA_ENABLE_INPUT_PICKER_CRITICAL_RND
+#include "input_pickers/critical_rnd.hh"
+#endif
