+++ src/solver/spot_fast_mode.hh
@@ -1,0 +1,8 @@
+#pragma once
+
+enum SPOT_FAST_T : char {
+  SPOT_FAST_OFF = 0,
+  SPOT_FAST_DET = 1,
+  SPOT_FAST_GFG_DECISION = 2,
+  SPOT_FAST_DET_AND_GFG = 3,
+};
