+++ src/solver/spot_nba_fastpath.hh
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "configuration.hh"
+#include "solver/spot_fast_mode.hh"
 #include "utils/verbose.hh"
 
 #include <algorithm>
@@ -18,13 +19,6 @@
 #include <spot/twaalgos/game.hh>
 #include <spot/twaalgos/mealy_machine.hh>
 #include <spot/twaalgos/synthesis.hh>
-
-enum SPOT_FAST_T : char {
-  SPOT_FAST_OFF = 0,
-  SPOT_FAST_DET = 1,
-  SPOT_FAST_GFG_DECISION = 2,
-  SPOT_FAST_DET_AND_GFG = 3,
-};
 
 namespace acacia::spot_fastpath {
